+++ src/hal/stm32h7/hal_crc.c
@@ -50,7 +50,7 @@ uint16_t crc_u16_start( void ){
 }
 
 
-uint16_t crc_u16_block( uint8_t *ptr, uint16_t length ){
+uint16_t crc_u16_block( uint8_t *ptr, uint32_t length ){
 
     crc_u16_start();
 
@@ -61,7 +61,7 @@ uint16_t crc_u16_block( uint8_t *ptr, uint16_t length ){
 
 // compute the CRC for the given block with initialization or
 // finishing the last two 0 bytes
-uint16_t crc_u16_partial_block( uint16_t crc, uint8_t *ptr, uint16_t length ){
+uint16_t crc_u16_partial_block( uint16_t crc, uint8_t *ptr, uint32_t length ){
 
     return HAL_CRC_Accumulate( &hcrc, (uint32_t *)ptr, length );
 }
